requirements.txt
opencv-python>=4.5.0
tensorflow>=2.5.0
keras>=2.4.0
numpy>=1.19.0
pillow>=8.0.0
dlib>=19.22.0
imutils>=0.5.4
h5py>=3.1.0
scikit-learn>=0.24.0
tqdm>=4.60.0